% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_coefs.R
\name{boot_betas}
\alias{boot_betas}
\title{boot_betas}
\usage{
boot_betas(data, idx)
}
\arguments{
\item{data}{User table with predictor variable in 1st colum and response variable in 2nd}

\item{idx}{indices to use for generation of bootstrap samples}
}
\value{
estimated coefficients using simple linear regression
}
\description{
Function to use with boot function to obtain bootstrap estimates of linear regression model coefficients
To use boot function need data and indices
}
\examples{
\dontrun{
boot_betas(data,idx)
  }

}
\keyword{internal}
