% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{Correlation}
\alias{Correlation}
\title{Correlation}
\usage{
Correlation(x, y, x_lab, y_lab)
}
\arguments{
\item{x}{vector of predictor variable data}

\item{y}{vector of response variable data}

\item{x_lab}{predictor variable column title}

\item{y_lab}{response variable column title}
}
\value{
Pearson or Spearman's correlation constant with p values
}
\description{
This function checks if data is appropriate for calculation of a correlation coefficient
If appropriate, Pearson's (if linear and normal distribution) or Spearman's (linear relationship, not normal distribution)
}
\examples{
\dontrun{
Correlation(x,y,Wages,Travel)
}

}
\keyword{internal}
