% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Introduction.R
\name{intro}
\alias{intro}
\title{intro}
\usage{
intro()
}
\value{
path to users data
}
\description{
This function helps user to determine if their data is appropriate for simple linear regression
}
\examples{
\dontrun{
intro()
}

}
\keyword{internal}
