% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{Load}
\alias{Load}
\title{Load
This function will load csv,txt and excel files based on extension}
\usage{
Load(loc)
}
\arguments{
\item{loc}{path to users data file}
}
\value{
table with loaded user data
}
\description{
Load
This function will load csv,txt and excel files based on extension
}
\examples{
\dontrun{
Load("C:\user\file.txt")
}

}
\keyword{internal}
