% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputLM.R
\name{SimpleLinear}
\alias{SimpleLinear}
\title{Simple Linear}
\usage{
SimpleLinear(x, y)
}
\arguments{
\item{x}{vector of predictor variable data points}

\item{y}{vector or response variable data points}
}
\value{
summary of linear model
}
\description{
This function performs simple linear regression on appropriate data
This function also provides an explanation of the output
}
\examples{
\dontrun{
SimpleLinear(x,y)
}

}
\keyword{internal}
