% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM101.R
\name{SLR101}
\alias{SLR101}
\title{LM101}
\usage{
SLR101()
}
\value{
Linear model and confidence intervals
}
\description{
This is an interactive function guiding user through creation of a model using simple linear regression or bootstrapping.
Confidence intervals for predictions are included as well.
}
\examples{
 \dontrun{
 initiating function

 SLR101()
 }

}
