% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{transformation}
\alias{transformation}
\title{transformation}
\usage{
transformation(x, y, x_lab, y_lab)
}
\arguments{
\item{x}{a vector or predictor variable data points}

\item{y}{a vector of response variable data points}

\item{x_lab}{the title o the predictor column}

\item{y_lab}{title of the response column}
}
\value{
transformed predictor and/or response data or NA
}
\description{
This function performs box cox transformation on data that does not have a normal distribution
After transformation, a scatterplot is done to check if the two variables have a linear relationship
}
\examples{
\dontrun{
transformation(x,y,x_lab,l_lab)
}

}
\keyword{internal}
