% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheck.R
\name{validity}
\alias{validity}
\title{validity}
\usage{
validity(x, y, x_lab, y_lab)
}
\arguments{
\item{x}{vector of predictor variable data}

\item{y}{vector of response variable data}

\item{x_lab}{title of the predictor variable column}

\item{y_lab}{title of the response variable column}
}
\value{
a list containing the original x and y vectors, transformed x and/or y vectors or NA,NA
}
\description{
This function checks that a previously generated linear model follows the assumptions of simple linear regression
If appropriate and necessary the data are transformed using box cox transformation
}
\examples{
\dontrun{
validity(x,y,Wages, Travel)
}

}
\keyword{internal}
