% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualization.R
\name{Visual}
\alias{Visual}
\title{Visual}
\usage{
Visual(x, y, x_lab, y_lab, p = 19, c = "blue")
}
\arguments{
\item{x}{predictor variable data}

\item{y}{response variable data}

\item{x_lab}{predictor variable column title}

\item{y_lab}{response variable column title}

\item{p}{is pch and defaults to 19}

\item{c}{is color of data and defaults to blue}
}
\value{
pdf containing plots and outcomes of shapiro wilks test
}
\description{
This function allows user to visualize data with scatter plots, histograms and density plots
This function is also designed to help user determine if their data has a normal distribution
}
\examples{
\dontrun{
Visual(x,y,Wages, Travel)
}
}
\keyword{internal}
