% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVectorsTitles.R
\name{xyVectors}
\alias{xyVectors}
\title{xyVectors}
\usage{
xyVectors(table)
}
\arguments{
\item{table}{loaded user table}
}
\value{
list containing predictor and response variables and their column titles
}
\description{
This function separates loaded table into predictor and response variable vectors
}
\examples{
\dontrun{
xyVectors(table)
}

}
\keyword{internal}
